+++ src/Libraries/SmartStore.Core/Domain/Payments/PaymentMethod.cs
@@ -1,5 +1,6 @@
 ﻿using System.Runtime.Serialization;
 using SmartStore.Core.Domain.Localization;
+using SmartStore.Core.Domain.Stores;
 
 namespace SmartStore.Core.Domain.Payments
 {
@@ -7,7 +8,7 @@ namespace SmartStore.Core.Domain.Payments
 	/// Represents a payment method
 	/// </summary>
 	[DataContract]
-	public partial class PaymentMethod : BaseEntity, ILocalizedEntity
+	public partial class PaymentMethod : BaseEntity, ILocalizedEntity, IStoreMappingSupported
 	{
 		/// <summary>
 		/// Gets or sets the payment method system name
@@ -27,5 +28,11 @@ namespace SmartStore.Core.Domain.Payments
         /// <see cref="https://en.wikipedia.org/wiki/Cash_rounding"/>
         [DataMember]
         public bool RoundOrderTotalEnabled { get; set; }
-    }
+
+		/// <summary>
+		/// Gets or sets a value indicating whether the entity is limited/restricted to certain stores
+		/// </summary>
+		[DataMember]
+		public bool LimitedToStores { get; set; }
+	}
 }