+++ src/Libraries/SmartStore.Services/Shipping/ShippingExtentions.cs
@@ -5,7 +5,7 @@ using SmartStore.Core.Plugins;
 
 namespace SmartStore.Services.Shipping
 {
-    public static class ShippingExtentions
+	public static class ShippingExtentions
     {
         public static bool IsShippingRateComputationMethodActive(this Provider<IShippingRateComputationMethod> srcm, ShippingSettings shippingSettings)
         {
@@ -22,15 +22,6 @@ namespace SmartStore.Services.Shipping
 				return false;
 
 			return shippingSettings.ActiveShippingRateComputationMethodSystemNames.Contains(srcm.Metadata.SystemName, StringComparer.OrdinalIgnoreCase);
-        }
-
-        public static bool CountryRestrictionExists(this ShippingMethod shippingMethod, int countryId)
-        {
-            if (shippingMethod == null)
-                throw new ArgumentNullException("shippingMethod");
-
-            bool result = shippingMethod.RestrictedCountries.ToList().Find(c => c.Id == countryId) != null;
-            return result;
         }
     }
 }