+++ src/Presentation/SmartStore.Web/Administration/Models/Payments/PaymentMethodEditModel.cs
@@ -6,7 +6,7 @@ using SmartStore.Web.Framework.Modelling;
 
 namespace SmartStore.Admin.Models.Payments
 {
-	public class PaymentMethodEditModel : TabbableModel, ILocalizedModel<PaymentMethodLocalizedModel>
+	public class PaymentMethodEditModel : TabbableModel, ILocalizedModel<PaymentMethodLocalizedModel>, IStoreSelector
 	{
 		public PaymentMethodEditModel()
 		{
@@ -34,7 +34,12 @@ namespace SmartStore.Admin.Models.Payments
 
         [SmartResourceDisplayName("Admin.Configuration.Payment.Methods.RoundOrderTotalEnabled")]
         public bool RoundOrderTotalEnabled { get; set; }
-    }
+
+		// Store mapping
+		public bool LimitedToStores { get; set; }
+		public IEnumerable<SelectListItem> AvailableStores { get; set; }
+		public int[] SelectedStoreIds { get; set; }
+	}
 
 
 	public class PaymentMethodLocalizedModel : ILocalizedModelLocal